+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.0.0-alpha44
+
+## Features
+
+* Make select, multiselect + boolean callbacks abstract
+* Refactoring for DI integrations
+
+## Bugfixes
+
+* None
+
 # Version 1.0.0-alpha43
 
 ## Features